api_yamdb/api/v_1/filters.py
from django_filters import rest_framework as filters

from reviews.models import Title


class TitlesFilter(filters.FilterSet):
    genre = filters.CharFilter(
        field_name='genre__slug',
        lookup_expr='iexact'
    )

    category = filters.CharFilter(
        field_name='category__slug',
        lookup_expr='iexact'
    )

    name = filters.CharFilter(
        field_name='name',
        lookup_expr='icontains'
    )

    class Meta:
        model = Title
        fields = ('year',)
